Move types from lib.rs into their own submodules
